PR 41 - Correções e Melhorias.
