+++ api/InOut.API/Controllers/BrandController.cs
@@ -1,19 +1,22 @@
-﻿using InOut.Domain.Exceptions;
+﻿using InOut.API.Builders;
+using InOut.Domain.Exceptions;
 using InOut.Domain.Models.Brand;
+using InOut.Infrastructure.Repositories.Interfaces;
 using InOut.Service.Services.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 
 namespace InOut.API.Controllers
 {
     [ApiController]
-    //[Route("/api/v1/brands")]
     public class BrandController : ControllerBase
     {
         private readonly IBrandService _brandService;
+        private readonly IBrandRepository _brandRepository;
 
-        public BrandController(IBrandService brandService)
+        public BrandController(IBrandService brandService, IBrandRepository brandRepository)
         {
             _brandService = brandService;
+            _brandRepository = brandRepository;
         }
 
         [HttpPost]
@@ -49,6 +52,23 @@ namespace InOut.API.Controllers
             catch (Exception ex)
             {
                 throw;
+            }
+        }
+
+        [HttpGet]
+        [Route("/api/v1/brands/getAllBrandsInfo")]
+        public async Task<IActionResult> GetAllBrandInfo()
+        {
+            try
+            {
+                var brands = _brandRepository.GetAllAsNoTracking();
+
+                return Ok(brands);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new ResponseModelBuilder().WithMessage("Brands not found: " + ex.Message)
+                                                            .Build());
             }
         }
     }
