+++ api/InOut.API/Controllers/ProductController.cs
@@ -1,5 +1,7 @@
-﻿using InOut.Domain.Exceptions;
+﻿using InOut.API.Builders;
+using InOut.Domain.Exceptions;
 using InOut.Domain.Models.Product;
+using InOut.Infrastructure.Repositories.Interfaces;
 using InOut.Service.Services.Interfaces;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -7,14 +9,15 @@ using Microsoft.AspNetCore.Mvc;
 namespace InOut.API.Controllers
 {
     [ApiController]
-    //[Route("/api/v1/products")]
     public class ProductController : ControllerBase
     {
         private readonly IProductService _productService;
+        private readonly IProductRepository _productRepository;
 
-        public ProductController(IProductService productService)
+        public ProductController(IProductService productService, IProductRepository productRepository)
         {
             _productService = productService;
+            _productRepository = productRepository;
         }
 
         [HttpPost]
@@ -51,6 +54,23 @@ namespace InOut.API.Controllers
             catch (Exception ex)
             {
                 throw;
+            }
+        }
+
+        [HttpGet]
+        [Route("/api/v1/products/getAllProductsInfo")]
+        public async Task<IActionResult> GetAllProductsInfo()
+        {
+            try
+            {
+                var products = _productRepository.GetAllAsNoTracking();
+
+                return Ok(products);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(new ResponseModelBuilder().WithMessage("Products not found: " + ex.Message)
+                                                            .Build());
             }
         }
     }
